requirements.txt
slack-bolt>=1.18.0,<2.0
slack-sdk>=3.21.0,<4.0
openai>=1.0.0,<2.0
httpx>=0.25.0,<1.0
python-dotenv>=1.0.0,<2.0
